Chanigng to pointer ot int

+++ 0x0B-more_malloc_free/0-malloc_checked.c
@@ -8,7 +8,7 @@
  */
 void *malloc_checked(unsigned int b)
 {
-	float *pointer;
+	int *pointer;
 
 	pointer = malloc(b);
 	if (pointer == NULL)
